backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
dj-database-url==2.2.0
celery==5.4.0
redis==5.0.7
django-celery-beat==2.6.0
python-dotenv==1.0.1
gunicorn==22.0.0
whitenoise==6.7.0
